# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Plumbing Business AI Platform API
  description: |
    Comprehensive API for the production-ready plumbing business AI platform.
    
    This API provides endpoints for managing customers, jobs, conversations, AI interactions,
    emergency responses, scheduling, pricing, warranties, and system monitoring.
    
    ## Features
    - **Customer Management**: Complete customer lifecycle management
    - **AI-Powered Conversations**: Intelligent message processing and response generation
    - **Emergency Response**: Real-time emergency detection and routing
    - **Dynamic Pricing**: Intelligent pricing based on multiple factors
    - **Scheduling Optimization**: Advanced technician scheduling and route optimization
    - **Warranty Management**: Comprehensive warranty tracking and claims processing
    - **Performance Monitoring**: Real-time system health and performance metrics
    
    ## Authentication
    All endpoints require JWT-based authentication unless otherwise specified.
    Include the token in the Authorization header: `Bearer {token}`
    
    ## Rate Limiting
    API requests are rate-limited to prevent abuse:
    - Standard endpoints: 1000 requests per hour per user
    - Search endpoints: 100 requests per minute per user
    - Webhook endpoints: 10,000 requests per hour
    
    ## Error Handling
    The API uses standard HTTP status codes and returns structured error responses:
    ```json
    {
      "error": {
...
        }
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /customers:
    get:
      tags: [Customers]
      summary: List customers
      description: Retrieve paginated list of customers with optional filtering
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 25
        - name: search
          in: query
          schema:
            type: string
          description: Search in customer name, email, or phone
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, blocked]
        - name: customerType
          in: query
          schema:
            type: string
            enum: [residential, commercial, property_manager]
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [name, createdAt, totalRevenue, lastServiceDate]
            default: createdAt
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: Customers retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  customers:
                    type: array
                    items:
                      $ref: '#/components/schemas/Customer'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
                  totalCount:
                    type: integer
                    example: 150
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Customers]
      summary: Create customer
      description: Create a new customer profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerCreate'
      responses:
        '201':
          description: Customer created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  customer:
                    $ref: '#/components/schemas/Customer'
                  message:
                    type: string
                    example: Customer created successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /jobs:
    get:
      tags: [Jobs]
      summary: List jobs
      description: Retrieve paginated list of jobs with filtering options
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 25
        - name: status
          in: query
          schema:
            type: string
            enum: [scheduled, assigned, in_progress, completed, cancelled]
        - name: priority
          in: query
          schema:
            type: string
            enum: [low, normal, high, emergency]
        - name: serviceType
          in: query
          schema:
            type: string
        - name: customerId
          in: query
          schema:
            type: string
            format: uuid
        - name: technicianId
          in: query
          schema:
            type: string
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date-time
        - name: dateTo
          in: query
          schema:
            type: string
            format: date-time
        - name: include
          in: query
          schema:
            type: string
          description: 'Related data to include: customer,technician,quote,materials'
      responses:
        '200':
          description: Jobs retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  jobs:
                    type: array
                    items:
                      $ref: '#/components/schemas/Job'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
                  summary:
                    type: object
                    properties:
                      totalJobs:
                        type: integer
                      byStatus:
                        type: object
                        additionalProperties:
                          type: integer
                      totalRevenue:
                        type: number

    post:
      tags: [Jobs]
      summary: Create job
      description: Create a new service job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobCreate'
      responses:
        '201':
          description: Job created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  job:
                    $ref: '#/components/schemas/Job'

  # AI Services Endpoints
...
  /webhooks/google-voice:
    post:
      tags: [Webhooks]
      summary: Google Voice webhook
      description: Handle incoming messages and calls from Google Voice
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message, eventType]
              properties:
                message:
                  type: object
                  properties:
                    text:
                      type: string
                    phoneNumber:
                      type: string
                    timestamp:
                      type: string
                      format: date-time
                    type:
                      type: string
                      enum: [sms, voice_call, voicemail]
                    direction:
                      type: string
                      enum: [inbound, outbound]
                eventType:
                  type: string
                  enum: [message_received, call_received, voicemail_received]
      responses:
        '200':
          description: Webhook processed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  processed:
                    type: boolean
                  conversationId:
                    type: string
                    format: uuid
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Customer:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        zipCode:
          type: string
        customerType:
          type: string
          enum: [residential, commercial, property_manager]
        status:
          type: string
          enum: [active, inactive, blocked]
        totalRevenue:
          type: number
          minimum: 0
        totalServices:
          type: integer
          minimum: 0
        averageRating:
          type: number
          minimum: 0
          maximum: 5
        firstServiceDate:
          type: string
          format: date-time
        lastServiceDate:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CustomerCreate:
      type: object
      required: [name, phone]
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        phone:
          type: string
          pattern: '^\+1[0-9]{10}$'
        address:
          type: string
        city:
          type: string
        state:
          type: string
        zipCode:
          type: string
        customerType:
          type: string
          enum: [residential, commercial, property_manager]
          default: residential
        notes:
          type: string
        source:
          type: string
          enum: [website, phone, referral, google_voice, walk_in]
...
    Job:
      type: object
      properties:
        id:
          type: string
          format: uuid
        customerId:
          type: string
          format: uuid
        type:
          type: string
          enum: [estimate, service, maintenance, emergency_repair, warranty_service]
        serviceType:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [scheduled, assigned, in_progress, completed, cancelled]
        priority:
          type: string
          enum: [low, normal, high, emergency]
        scheduledDate:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        estimatedDuration:
          type: integer
          description: Duration in minutes
        actualDuration:
          type: integer
          description: Actual duration in minutes
        assignedTechnicianId:
          type: string
        totalCost:
          type: number
          minimum: 0
        customerRating:
          type: integer
          minimum: 1
          maximum: 5
        warrantyPeriodMonths:
          type: integer
          minimum: 0
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    JobCreate:
      type: object
      required: [customerId, type, serviceType, description]
      properties:
        customerId:
          type: string
          format: uuid
        type:
          type: string
          enum: [estimate, service, maintenance, emergency_repair, warranty_service]
        serviceType:
          type: string
        description:
          type: string
          minLength: 10
        priority:
          type: string
          enum: [low, normal, high, emergency]
          default: normal
        scheduledDate:
          type: string
          format: date-time
        estimatedDuration:
          type: integer
          minimum: 15
          maximum: 480
        assignedTechnicianId:
          type: string
        notes:
          type: string
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
        limit:
          type: integer
          minimum: 1
        totalPages:
          type: integer
          minimum: 1
        hasNext:
          type: boolean
        hasPrev:
          type: boolean
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Invalid input data
            details:
              type: object
              additionalProperties: true
  responses:
    BadRequest:
      description: Bad request - Invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized - Invalid or missing token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    Conflict:
      description: Conflict - Resource already exists
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...